+++ webots_ros2_universal_robot/CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog for package webots_ros2_universal_robot
 ------------------
 * Added an 'universal_robot_multiple' simulation and launch file.
 * Added an 'universal_robot_rviz' simulation and launch file.
+* Added support for the goal tolerance in the action server.
 * Fixed the action server that was never reaching the `SUCCESSFUL` state.
 
 0.0.2 (2019-09-23)